requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
supabase==2.28.3
pydantic-settings==2.14.0
anthropic==0.96.0
httpx==0.28.1
pyjwt==2.10.1
bcrypt==4.2.1
cryptography==41.0.7
python-multipart==0.0.20
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
